requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.129.0
h11==0.16.0
httptools==0.7.1
idna==3.11
mysql-connector-python==9.6.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
starlette==0.52.1
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
flet==0.82.2